fix: unwrap Annotated and propagate whole-subject captures through OR patterns
_is_fact_typed_annotation()'s generic fallback recursed into the
subscript's own head (annotation.value), correct for plain Fact[int] but
wrong for Annotated[Fact[int], metadata] (PEP 593) -- the real type lives
in the slice's first element, not the head. Fixed with a dedicated
Annotated branch mirroring the existing Optional/Union handling.

The ast.Match branch's whole-subject-capture recognition had no
counterpart for an OR pattern where every alternative independently
captures the whole subject under the same name. Python only requires
every alternative to bind the same *names*, not the same binding shape,
so name consistency alone isn't sufficient evidence -- fixed by requiring
every alternative to itself be exactly ast.MatchAs with the identical
name before trusting it as a whole-subject alias.

Split tests/test_fact_detector_misuse_alias_edge_cases.py out of
test_fact_detector_misuse_def_time_scope.py (which reached 1198/1200
lines) -- mechanical extraction of the four most recent alias-resolution
test classes, mirroring that file's own earlier split.
